backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fugle-marketdata
greenlet==3.2.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jwt==1.3.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==8.3.0
numpy==2.2.5
pandas==2.2.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
